app/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
pydantic-settings==2.13.1
openai==2.24.0
redis==7.2.1
SQLAlchemy==2.0.47
asyncpg==0.31.0